Collected all parameters in structured definitions
All the parameters are now structured with shared interfaces. This is
the first step to allowing the UI to properly access the parameters of
the different protections and to verify the project files to locate
wrong parameters.

One big change was done. The "rename public" parameter for the renamer
was dropped, because this behaviour can easily be produced by properly
setting the rules.