Fixed handling of Windows Forms DataGridColumns
The compatibility mode for Windows Forms not recognises the
DataPropertyName of DataGridColumn and avoids renaming the properties
referenced there.

Also fixed the handling of the attributes in case the values are
duplicated on the evaluation stack.

Another fix for handling Bindings in Windows Forms was added also. The
binding constructor is now properly handled.