refs #2 Updated LZMA Implementation to 18.05
As part of this, the runtime implementation for decompressing LZMA was
also updated. Now it is directly referencing the original source files
and is not a version that is assembled by a lot of copy & paste.

The problem now is, that the LZMA runtime consists of multiple classes
that depend on each other. That is something the inject helper is not
able to handle right now.

While this commit does not work, it is the baseline for updating the
inject helper for it to handle and resolve more complex dependencies.