refs #2 Resources Protection with new inject
The Resource Protection is now using the new code injection. Some
changes to the CILCodeGen of the dynamic cipher were required to avoid
a NullReferenceException because the cipher injection was done before
the method was linked to the target type (because the type is not
injected at this point for sure).